requirements.txt
numpy==1.16.0
opencv_python==4.2.0.32
scipy==1.2.0
Pillow==8.1.0
tensorboardX==2.0
